requirements.txt
PySide6>=6.5.0,<7.0
Pillow>=10.0.0
exifread>=3.0.0
shapely>=2.0.0
rtree>=1.0.0
